Raise error if MFA code is incorrect
Previously if someone typed the MFA code incorrectly, aws-mfa would store
invalid data in the local credentials cache. The changes in this commit check if the
call to fetch credentials from AWS succeeded before storing the result. It
now throws an error if the MFA code is incorrect and the cache is no longer
corrupted.